requirements.txt
brotlipy==0.7.0
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
cffi @ file:///C:/b/abs_49n3v2hyhr/croot/cffi_1670423218144/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
cloudpickle==2.2.1
cryptography @ file:///C:/b/abs_8ecplyc3n2/croot/cryptography_1677533105000/work
cycler==0.11.0
ess-python-tools==2.14.1
fonttools==4.38.0
gym==0.26.2
gym-notices==0.0.8
h5py==3.8.0
idna @ file:///C:/b/abs_bdhbebrioa/croot/idna_1666125572046/work
importlib-metadata==6.6.0
kiwisolver==1.4.4
lxml==4.9.2
matplotlib==3.5.3
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186163140/work
mkl-service==2.4.0
numpy @ file:///C:/ci/numpy_and_numpy_base_1653574840943/work
packaging==23.1
Pillow==9.4.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyepics==3.5.1
pyOpenSSL @ file:///C:/b/abs_552w85x1jz/croot/pyopenssl_1677607703691/work
pyparsing==3.0.9
PySocks @ file:///C:/ci/pysocks_1594394709107/work
python-dateutil==2.8.2
requests @ file:///C:/ci/requests_1657735288441/work
scipy==1.7.3
six @ file:///tmp/build/80754af9/six_1644875935023/work
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
typing_extensions @ file:///C:/Windows/TEMP/abs_dd2d0moa85/croots/recipe/typing_extensions_1659638831135/work
urllib3 @ file:///C:/b/abs_9bcwxczrvm/croot/urllib3_1673575521331/work
webdavclient3==3.14.6
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306165655/work
wincertstore==0.2
zipp==3.15.0
